+++ packages/react/src/components/Dropzone/DropzoneContext.tsx
@@ -1,16 +1,9 @@
 import { createContext, useContext } from "react";
 
-import { WorkspaceElement } from "edifice-ts-client";
-
-import { Status } from "../../utils/Status";
-
 export interface DropzoneContextType {
   inputRef: React.MutableRefObject<HTMLInputElement | null>;
   importMessage?: string;
   files: File[];
-  status: Status | undefined;
-  handleDelete: (element: WorkspaceElement, index: number) => void;
-  handleSave: any;
   setFiles: any;
 }
