+++ packages/react/src/hooks/useDropzone/useDropzone.ts
@@ -1,12 +1,7 @@
 import { useCallback, useState } from "react";
 
-import { WorkspaceElement, odeServices } from "edifice-ts-client";
-
-import { Status } from "../../utils/Status";
-
 const useDropzone = (ref: any, onFilesChange: (files: any) => void) => {
   const [dragging, setDragging] = useState<boolean>(false);
-  const [status, setStatus] = useState<Status | undefined>("idle");
 
   const handleDragging = useCallback(
     (event: React.DragEvent<HTMLDivElement>) => {
@@ -33,38 +28,10 @@ const useDropzone = (ref: any, onFilesChange: (files: any) => void) => {
     }
   };
 
-  async function handleSave(file: File) {
-    setStatus("loading");
-
-    try {
-      const doc = await odeServices.workspace().saveFile(file);
-
-      if (!doc) throw new Error("No document found");
-
-      if (doc._id) {
-        setStatus("success");
-        return doc;
-      }
-    } catch (error) {
-      setStatus("error");
-      console.log(error);
-    }
-  }
-
-  async function handleDelete(element: WorkspaceElement) {
-    try {
-      await odeServices.workspace().deleteFile([element]);
-    } catch (error) {
-      console.log(error);
-    }
-  }
-
   return {
     handleDragging,
     handleDragLeave,
     handleDrop,
-    handleSave,
-    handleDelete,
     dragging,
     status,
   };
